requirements.txt
numpy
scikit-learn
spacy
transformers==4.36.2
sentence-transformers==2.3.1
keybert==0.8.0
langchain
langchain_groq
python-dotenv
groq
pandas
datasets
torch
accelerate
faiss-cpu
pymongo
requests
chromadb
langchain-community
google-search-results